requirements-chess.txt
python-chess>=1.10.0
requests>=2.28.0
tqdm>=4.64.0
stockfish>=3.28.0
# Dependencies for analysis script
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0 